+++ whatsapp/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][keepachangelog],
 and this project adheres to [Semantic Versioning][semver].
 
+## [1.0.4] - 2022-02-06
+
+### Added
+
+- Added new types of messages (location, attachment, buttons, sections)
+- Added event that is sent whenever a message is received
+
 ## [1.0.3] - 2022-02-04
 
 ### Changed